Fatih

+++ Backend/src/main/java/com/softwareProject/banksApplication/dto/request/invoice/InvoiceSaveRequest.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -14,6 +15,6 @@ import java.time.LocalDateTime;
 public class InvoiceSaveRequest {
     private int invoiceNo;
     private InvoiceInfo.InvoiceType invoiceType;
-    private int invoiceAmount;
+    private BigDecimal invoiceAmount;
     private boolean autobill;
 }
