+++ Backend/src/main/java/com/softwareProject/banksApplication/entity/InvoiceInfo.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -27,7 +28,7 @@ public class InvoiceInfo {
     private InvoiceType invoiceType;
 
     @Column(name = "invoice_amount")
-    private int invoiceAmount;
+    private BigDecimal invoiceAmount;
 
     @Column(name = "pay_date")
     private LocalDateTime payDate;
